refactor: move createMosaicArrangement into EditorMosaic
